shared/pyproject.toml
[tool.poetry]
name = "rag-shared"
version = "0.1.0"
description = "Shared FAISS embedder and vector store library"
authors = ["kurodayu23"]
packages = [{include = "shared"}]

[tool.poetry.dependencies]
python = "^3.11"
faiss-cpu = "^1.7.4"
numpy = "^1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
